Scale the score difference
Scale the score difference: Consider scaling the Elo change based on the score difference. If a game was very close (e.g., 21-20), the change in Elo might be smaller than if the game were a blowout (e.g., 21-5). This system could be more reflective of player skill, as a higher-rated player winning by a slim margin against a lower-rated player may indicate that their skill levels are closer than their ratings suggest.